requirements.txt
flask
pandas
sqlalchemy
pymysql
faker
python-dotenv